+++ Services/ServiceConfiguration.cs
@@ -26,6 +26,7 @@ namespace SwitchBlade.Services
             services.AddSingleton<ISettingsService>(sp => sp.GetRequiredService<SettingsService>());
             services.AddSingleton<ThemeService>();
             services.AddSingleton<IDispatcherService, WpfDispatcherService>();
+            services.AddSingleton<IIconService, IconService>();
 
             // Logger
             services.AddSingleton<ILogger>(Logger.Instance);
@@ -53,7 +54,8 @@ namespace SwitchBlade.Services
                 return new MainViewModel(
                     providers,
                     sp.GetRequiredService<ISettingsService>(),
-                    sp.GetRequiredService<IDispatcherService>()
+                    sp.GetRequiredService<IDispatcherService>(),
+                    sp.GetRequiredService<IIconService>()
                 );
             });